Refactor Bucket requests objects to one file.
Creating one file per request type was getting tedious. This is
in preparation of adding more request types for #821.
