+++ google/cloud/storage/storage_client.bzl
@@ -24,12 +24,11 @@ storage_client_HDRS = [
     "internal/empty_response.h",
     "internal/generic_object_request.h",
     "internal/generic_request.h",
-    "internal/get_bucket_metadata_request.h",
+    "internal/bucket_requests.h",
     "internal/get_object_metadata_request.h",
     "internal/google_application_default_credentials_file.h",
     "internal/http_response.h",
     "internal/insert_object_media_request.h",
-    "internal/list_buckets_request.h",
     "internal/list_object_acl_request.h",
     "internal/list_objects_request.h",
     "internal/logging_client.h",
@@ -77,11 +76,10 @@ storage_client_SRCS = [
     "internal/curl_streambuf.cc",
     "internal/delete_object_request.cc",
     "internal/empty_response.cc",
-    "internal/get_bucket_metadata_request.cc",
+    "internal/bucket_requests.cc",
     "internal/get_object_metadata_request.cc",
     "internal/google_application_default_credentials_file.cc",
     "internal/insert_object_media_request.cc",
-    "internal/list_buckets_request.cc",
     "internal/list_object_acl_request.cc",
     "internal/list_objects_request.cc",
     "internal/logging_client.cc",
